refactor: decompose onboarding-dialog god file into focused modules
- Extract each step into apps/web/src/components/onboarding/steps/
  (welcome, profile, provider, memory, success)
- Extract state management into useOnboardingState hook
- Slim onboarding-dialog.tsx to a thin step-switching orchestrator (~54 lines)
- Eliminate ~217 lines of duplicated provider config logic by extending
  shared ProviderConfig with optional saveLabel/onSaved/showDisconnect props
- Replace hardcoded emerald color with semantic success token in success step
