+++ apps/web/src/components/onboarding/onboarding-dialog.tsx
@@ -1,644 +1,16 @@
-import { ArrowLeftIcon, CheckCircle2Icon, SparklesIcon, PlusIcon } from 'lucide-react';
-import * as React from 'react';
+import { MemoryStep } from './steps/memory-step';
+import { ProfileStep } from './steps/profile-step';
+import { ProviderStep } from './steps/provider-step';
+import { SuccessStep } from './steps/success-step';
+import { WelcomeStep } from './steps/welcome-step';
+import { useOnboardingState } from './use-onboarding-state';
 
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
-
-import { PROVIDER_META } from '@stitch/shared/providers/catalog';
-import { PROVIDER_IDS, type ProviderId } from '@stitch/shared/providers/types';
-
-import { FieldGroup, NoFieldsNote } from '@/components/settings/providers/field-group';
-import { ProviderLogo } from '@/components/settings/providers/provider-logo';
-import {
-  buildProviderConfigBody,
-  hydrateProviderConfigState,
-  resolveDefaultAuthMethod,
-  type FieldValues,
-} from '@/components/settings/providers/utils';
-import { Button } from '@/components/ui/button';
 import { Dialog, DialogContent, DialogHeader, DialogTitle } from '@/components/ui/dialog';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
-import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import { useSaveProviderConfigMutation } from '@/lib/mutations/provider-config';
-import {
-  embeddingProviderModelsQueryOptions,
-  providerConfigQueryOptions,
-  providersQueryOptions,
-  type ProviderSummary,
-} from '@/lib/queries/providers';
-import { saveSettingMutationOptions, settingsQueryOptions } from '@/lib/queries/settings';
-
-type OnboardingStep = 'welcome' | 'profile' | 'provider' | 'memory' | 'success';
-
-const SUCCESS_CLOSE_DELAY_MS = 1200;
-const CURRENT_ONBOARDING_VERSION = '4';
-
-function getDetectedTimezone(): string {
-  const resolved = Intl.DateTimeFormat().resolvedOptions().timeZone;
-  return typeof resolved === 'string' && resolved.trim().length > 0 ? resolved.trim() : 'UTC';
-}
-
-function getTimezoneOptions(initialTimezone: string): string[] {
-  const intlWithSupportedValues = Intl as typeof Intl & {
-    supportedValuesOf?: (key: string) => string[];
-  };
-  const listed = intlWithSupportedValues.supportedValuesOf?.('timeZone') ?? [];
-  const preferred = [initialTimezone].filter((value) => value.length > 0);
-
-  if (listed.length === 0) {
-    return preferred;
-  }
-
-  return Array.from(new Set([...preferred, ...listed]));
-}
-
-function OnboardingProfileStep({
-  initialName,
-  initialTimezone,
-  isSaving,
-  onContinue,
-}: {
-  initialName: string;
-  initialTimezone: string;
-  isSaving: boolean;
-  onContinue: (name: string, timezone: string) => void;
-}) {
-  const detectedTimezone = React.useMemo(() => getDetectedTimezone(), []);
-  const [name, setName] = React.useState(initialName);
-  const [timezone, setTimezone] = React.useState(initialTimezone || detectedTimezone);
-  const [touched, setTouched] = React.useState(false);
-  const timezoneOptions = React.useMemo(
-    () => getTimezoneOptions(initialTimezone || detectedTimezone),
-    [detectedTimezone, initialTimezone],
-  );
-
-  const trimmed = name.trim();
-  const trimmedTimezone = timezone.trim();
-  const hasError = touched && trimmed.length === 0;
-  const hasTimezoneError = touched && trimmedTimezone.length === 0;
-
-  function handleSubmit(event: React.FormEvent<HTMLFormElement>) {
-    event.preventDefault();
-    setTouched(true);
-    if (trimmed.length === 0) {
-      return;
-    }
-    if (trimmedTimezone.length === 0) {
-      return;
-    }
-    onContinue(trimmed, trimmedTimezone);
-  }
-
-  return (
-    <form
-      className="mx-auto flex h-full w-full max-w-md flex-col justify-center gap-6"
-      onSubmit={handleSubmit}
-    >
-      <div className="space-y-2 text-center">
-        <h2 className="text-2xl font-semibold tracking-tight">Tell us your name</h2>
-        <p className="text-sm text-muted-foreground">
-          We&apos;ll use it to personalize responses and transcription speaker labels.
-        </p>
-      </div>
-
-      <div className="space-y-2">
-        <Label htmlFor="onboarding-name">Name</Label>
-        <Input
-          id="onboarding-name"
-          value={name}
-          onChange={(event) => setName(event.target.value)}
-          onBlur={() => setTouched(true)}
-          placeholder="Jane"
-          maxLength={80}
-          autoFocus
-        />
-        {hasError && <p className="text-xs text-destructive">Please enter your name.</p>}
-      </div>
-
-      <div className="space-y-2">
-        <Label htmlFor="onboarding-timezone">Timezone</Label>
-        <Select value={timezone} onValueChange={(value) => setTimezone(value ?? '')}>
-          <SelectTrigger id="onboarding-timezone" className="w-full">
-            <SelectValue placeholder="Select your timezone" />
-          </SelectTrigger>
-          <SelectContent className="max-h-80">
-            {timezoneOptions.map((value) => (
-              <SelectItem key={value} value={value}>
-                {value}
-              </SelectItem>
-            ))}
-          </SelectContent>
-        </Select>
-        {hasTimezoneError && <p className="text-xs text-destructive">Please select a timezone.</p>}
-      </div>
-
-      <Button
-        size="lg"
-        type="submit"
-        disabled={isSaving || trimmed.length === 0 || trimmedTimezone.length === 0}
-      >
-        {isSaving ? 'Saving...' : 'Continue'}
-      </Button>
-    </form>
-  );
-}
-
-function OnboardingProviderConfig({
-  provider,
-  onBack,
-  onConnected,
-}: {
-  provider: ProviderSummary;
-  onBack: () => void;
-  onConnected: () => void;
-}) {
-  const meta = (PROVIDER_IDS as readonly string[]).includes(provider.id)
-    ? PROVIDER_META[provider.id as ProviderId]
-    : undefined;
-  const enabledAuthMethods = React.useMemo(
-    () => meta?.authMethods.filter((method) => method.enabled) ?? [],
-    [meta],
-  );
-  const queryClient = useQueryClient();
-  const { data: existingConfig, isPending } = useQuery({
-    ...providerConfigQueryOptions(provider.id),
-    enabled: provider.enabled,
-  });
-
-  const existingMethod = (existingConfig?.auth as { method?: string } | undefined)?.method;
-  const defaultMethod = resolveDefaultAuthMethod(existingMethod, enabledAuthMethods);
-
-  const [activeTab, setActiveTab] = React.useState(defaultMethod);
-  const [fieldsByMethod, setFieldsByMethod] = React.useState<Record<string, FieldValues>>({});
-  const [extraFields, setExtraFields] = React.useState<FieldValues>({});
-  const [fieldErrors, setFieldErrors] = React.useState<Record<string, string>>({});
-
-  React.useEffect(() => {
-    if (!existingConfig || !meta) return;
-    const hydrated = hydrateProviderConfigState(
-      existingConfig as Record<string, unknown>,
-      enabledAuthMethods,
-    );
-    const activeMethod = hydrated.activeMethod;
-    if (activeMethod) {
-      setActiveTab(activeMethod);
-      setFieldsByMethod((prev) => ({ ...prev, [activeMethod]: hydrated.authFields }));
-    } else if (enabledAuthMethods[0]?.method) {
-      setActiveTab(enabledAuthMethods[0].method);
-    }
-    setExtraFields(hydrated.extraFields);
-  }, [enabledAuthMethods, existingConfig, meta]);
-
-  const saveMutation = useSaveProviderConfigMutation({
-    providerId: provider.id,
-    queryClient,
-    successMessage: `${meta?.displayName ?? 'Provider'} connected`,
-    errorMessage: 'Failed to save provider config',
-    onSuccess: () => {
-      setFieldsByMethod({});
-      setExtraFields({});
-      onConnected();
-    },
-  });
-
-  if (provider.enabled && isPending) {
-    return <div className="text-sm text-muted-foreground">Loading provider configuration...</div>;
-  }
-
-  if (!meta || enabledAuthMethods.length === 0) return null;
-
-  const currentMethodFields = fieldsByMethod[activeTab] ?? {};
-
-  function handleMethodFieldChange(key: string, value: string) {
-    setFieldsByMethod((prev) => ({
-      ...prev,
-      [activeTab]: { ...prev[activeTab], [key]: value },
-    }));
-    if (fieldErrors[key]) {
-      setFieldErrors((prev) => {
-        const next = { ...prev };
-        delete next[key];
-        return next;
-      });
-    }
-  }
-
-  function handleExtraFieldChange(key: string, value: string) {
-    setExtraFields((prev) => ({ ...prev, [key]: value }));
-    if (fieldErrors[key]) {
-      setFieldErrors((prev) => {
-        const next = { ...prev };
-        delete next[key];
-        return next;
-      });
-    }
-  }
-
-  function handleSave() {
-    if (!meta) return;
-
-    const errors: Record<string, string> = {};
-
-    for (const field of meta.extraFields) {
-      if (field.required && !extraFields[field.key]) {
-        errors[field.key] = `${field.label} is required`;
-      }
-    }
-
-    const methodDef = enabledAuthMethods.find((m) => m.method === activeTab);
-    if (methodDef) {
-      for (const field of methodDef.fields) {
-        if (field.required && !currentMethodFields[field.key]) {
-          errors[field.key] = `${field.label} is required`;
-        }
-      }
-    }
-
-    if (Object.keys(errors).length > 0) {
-      setFieldErrors(errors);
-      return;
-    }
-
-    setFieldErrors({});
-    const body = buildProviderConfigBody({
-      activeTab,
-      enabledAuthMethods,
-      currentMethodFields,
-      extraFields,
-      extraFieldDefs: meta.extraFields,
-    });
-    saveMutation.mutate(body);
-  }
-
-  function handleTabChange(value: string | null) {
-    if (value) setActiveTab(value);
-  }
-
-  const hasMultipleMethods = enabledAuthMethods.length > 1;
-  const activeMethodDef = enabledAuthMethods.find((m) => m.method === activeTab);
-
-  return (
-    <div className="flex h-full flex-col">
-      <div className="mb-6 flex items-center gap-2">
-        <Button variant="ghost" size="icon-sm" onClick={onBack} aria-label="Back to providers">
-          <ArrowLeftIcon className="size-4" />
-        </Button>
-        <div className="shrink-0 text-muted-foreground">
-          <ProviderLogo
-            providerId={provider.id}
-            providerName={meta.displayName}
-            className="size-5"
-          />
-        </div>
-        <div>
-          <h2 className="text-sm font-semibold">{meta.displayName}</h2>
-          <p className="text-xs text-muted-foreground">{provider.model_count} models available</p>
-        </div>
-      </div>
-
-      <div className="flex flex-1 flex-col gap-5">
-        {meta.extraFields.length > 0 && (
-          <FieldGroup
-            fields={meta.extraFields}
-            providerId={provider.id}
-            values={extraFields}
-            errors={fieldErrors}
-            onChange={handleExtraFieldChange}
-          />
-        )}
-
-        {hasMultipleMethods ? (
-          <Tabs value={activeTab} onValueChange={handleTabChange}>
-            <TabsList variant="line">
-              {enabledAuthMethods.map((m) => (
-                <TabsTrigger key={m.method} value={m.method}>
-                  {m.label}
-                </TabsTrigger>
-              ))}
-            </TabsList>
-            {enabledAuthMethods.map((m) => (
-              <TabsContent key={m.method} value={m.method} className="mt-4">
-                {m.fields.length > 0 ? (
-                  <FieldGroup
-                    fields={m.fields}
-                    providerId={`${provider.id}-${m.method}`}
-                    values={fieldsByMethod[m.method] ?? {}}
-                    errors={fieldErrors}
-                    onChange={(key, value) =>
-                      setFieldsByMethod((prev) => ({
-                        ...prev,
-                        [m.method]: { ...prev[m.method], [key]: value },
-                      }))
-                    }
-                  />
-                ) : (
-                  <NoFieldsNote method={m.method} />
-                )}
-              </TabsContent>
-            ))}
-          </Tabs>
-        ) : (
-          activeMethodDef &&
-          (activeMethodDef.fields.length > 0 ? (
-            <FieldGroup
-              fields={activeMethodDef.fields}
-              providerId={provider.id}
-              values={currentMethodFields}
-              errors={fieldErrors}
-              onChange={handleMethodFieldChange}
-            />
-          ) : (
-            <NoFieldsNote method={activeMethodDef.method} />
-          ))
-        )}
-
-        <div className="flex items-center gap-2 pt-1">
-          <Button onClick={handleSave} disabled={saveMutation.isPending} size="sm">
-            {saveMutation.isPending ? 'Saving...' : 'Save and continue'}
-          </Button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function OnboardingProviderStep({ onConnected }: { onConnected: () => void }) {
-  const { data: providers } = useQuery(providersQueryOptions);
-  const [selected, setSelected] = React.useState<ProviderSummary | null>(null);
-
-  const selectableProviders = React.useMemo(() => {
-    if (!providers) return [];
-    return providers.filter((provider) => {
-      if (provider.enabled) return false;
-      if (!(PROVIDER_IDS as readonly string[]).includes(provider.id)) return false;
-      const meta = PROVIDER_META[provider.id as ProviderId];
-      return meta.authMethods.some((method) => method.enabled);
-    });
-  }, [providers]);
-
-  if (!providers) {
-    return <div className="text-sm text-muted-foreground">Loading providers...</div>;
-  }
-
-  if (selected) {
-    return (
-      <OnboardingProviderConfig
-        provider={selected}
-        onBack={() => setSelected(null)}
-        onConnected={onConnected}
-      />
-    );
-  }
-
-  return (
-    <div className="flex flex-col gap-6">
-      <div>
-        <h2 className="text-lg font-semibold">Setup Provider</h2>
-        <p className="mt-1 text-sm text-muted-foreground">
-          Connect one provider to unlock models and start chatting.
-        </p>
-      </div>
-
-      <div className="flex flex-col overflow-hidden rounded-xl border border-border/60">
-        {selectableProviders.length === 0 ? (
-          <div className="px-4 py-3 text-sm text-muted-foreground">
-            All available providers are already connected.
-          </div>
-        ) : (
-          selectableProviders.map((provider) => {
-            const meta = PROVIDER_META[provider.id as ProviderId];
-            return (
-              <div
-                key={provider.id}
-                className="flex items-center justify-between border-b border-border/50 px-4 py-3 last:border-0"
-              >
-                <div className="flex min-w-0 items-center gap-3">
-                  <div className="shrink-0 text-muted-foreground">
-                    <ProviderLogo providerId={provider.id} providerName={meta.displayName} />
-                  </div>
-                  <div className="min-w-0">
-                    <p className="truncate text-sm font-medium">{meta.displayName}</p>
-                    {meta.description && (
-                      <p className="truncate text-xs text-muted-foreground">{meta.description}</p>
-                    )}
-                  </div>
-                </div>
-                <Button size="sm" variant="outline" onClick={() => setSelected(provider)}>
-                  <PlusIcon className="mr-1 size-3.5" />
-                  Connect
-                </Button>
-              </div>
-            );
-          })
-        )}
-      </div>
-    </div>
-  );
-}
-
-function OnboardingMemoryStep({
-  onComplete,
-  onBackToProviders,
-}: {
-  onComplete: () => void;
-  onBackToProviders: () => void;
-}) {
-  const queryClient = useQueryClient();
-  const { data: settings } = useQuery(settingsQueryOptions);
-  const { data: providerModels } = useQuery(embeddingProviderModelsQueryOptions);
-
-  const saveEnabled = useMutation(
-    saveSettingMutationOptions('memory.enabled', queryClient, { silent: true }),
-  );
-  const saveProvider = useMutation(
-    saveSettingMutationOptions('memory.embedding.providerId', queryClient, { silent: true }),
-  );
-  const saveModel = useMutation(
-    saveSettingMutationOptions('memory.embedding.modelId', queryClient, { silent: true }),
-  );
-
-  const modelOptions = React.useMemo(() => {
-    if (!providerModels)
-      return [] as Array<{ value: string; label: string; providerId: string; modelId: string }>;
-
-    const options: Array<{ value: string; label: string; providerId: string; modelId: string }> =
-      [];
-    for (const provider of providerModels) {
-      for (const model of provider.models) {
-        options.push({
-          value: `${provider.providerId}:${model.id}`,
-          label: `${provider.providerName} - ${model.name}`,
-          providerId: provider.providerId,
-          modelId: model.id,
-        });
-      }
-    }
-    return options;
-  }, [providerModels]);
-
-  const [selectedValue, setSelectedValue] = React.useState<string>('');
-
-  React.useEffect(() => {
-    if (!settings || modelOptions.length === 0) return;
-
-    const existingValue = `${settings['memory.embedding.providerId']}:${settings['memory.embedding.modelId']}`;
-    const hasExisting = modelOptions.some((option) => option.value === existingValue);
-    if (hasExisting) {
-      setSelectedValue(existingValue);
-      return;
-    }
-
-    setSelectedValue(modelOptions[0].value);
-  }, [modelOptions, settings]);
-
-  if (!settings || !providerModels) {
-    return <div className="text-sm text-muted-foreground">Loading memory settings...</div>;
-  }
-
-  const hasModels = modelOptions.length > 0;
-  const selectedOption = modelOptions.find((option) => option.value === selectedValue);
-  const isSaving = saveEnabled.isPending || saveProvider.isPending || saveModel.isPending;
-
-  function handleDisableMemories() {
-    void saveEnabled
-      .mutateAsync('false')
-      .then(onComplete)
-      .catch(() => undefined);
-  }
-
-  function handleEnableMemories() {
-    if (!selectedOption) return;
-
-    void Promise.all([
-      saveProvider.mutateAsync(selectedOption.providerId),
-      saveModel.mutateAsync(selectedOption.modelId),
-      saveEnabled.mutateAsync('true'),
-    ])
-      .then(onComplete)
-      .catch(() => undefined);
-  }
-
-  return (
-    <div className="mx-auto flex h-full w-full max-w-lg flex-col justify-center gap-6">
-      <div className="space-y-2 text-center">
-        <h2 className="text-2xl font-semibold tracking-tight">Enable memories?</h2>
-        <p className="text-sm text-muted-foreground">
-          Memories help Stitch remember preferences and recurring context across sessions.
-        </p>
-      </div>
-
-      {hasModels ? (
-        <div className="space-y-2">
-          <Label htmlFor="onboarding-memory-model">Embedding model</Label>
-          <Select value={selectedValue} onValueChange={(value) => setSelectedValue(value ?? '')}>
-            <SelectTrigger id="onboarding-memory-model" className="w-full">
-              <SelectValue placeholder="Select an embedding model">
-                {selectedOption?.label}
-              </SelectValue>
-            </SelectTrigger>
-            <SelectContent className="max-h-80">
-              {modelOptions.map((option) => (
-                <SelectItem key={option.value} value={option.value}>
-                  {option.label}
-                </SelectItem>
-              ))}
-            </SelectContent>
-          </Select>
-        </div>
-      ) : (
-        <p className="rounded-md border border-border/60 bg-muted/30 px-3 py-2 text-sm text-muted-foreground">
-          No embedding models in providers configured. Please add another provider that has one
-        </p>
-      )}
-
-      <div className="flex items-center justify-center gap-2">
-        {!hasModels && (
-          <Button variant="outline" onClick={onBackToProviders} disabled={isSaving}>
-            Add provider
-          </Button>
-        )}
-        <Button variant="outline" onClick={handleDisableMemories} disabled={isSaving}>
-          Not now
-        </Button>
-        <Button onClick={handleEnableMemories} disabled={isSaving || !hasModels || !selectedOption}>
-          {isSaving ? 'Saving...' : 'Enable memories'}
-        </Button>
-      </div>
-    </div>
-  );
-}
 
 export function OnboardingDialog() {
-  const queryClient = useQueryClient();
-  const { data: settings, isPending: isSettingsPending } = useQuery(settingsQueryOptions);
-  const { data: providers, isPending: isProvidersPending } = useQuery(providersQueryOptions);
+  const state = useOnboardingState();
 
-  const [step, setStep] = React.useState<OnboardingStep>('welcome');
-  const [dismissed, setDismissed] = React.useState(false);
-
-  const saveOnboardingStatus = useMutation(
-    saveSettingMutationOptions('onboarding.status', queryClient, { silent: true }),
-  );
-  const saveOnboardingVersion = useMutation(
-    saveSettingMutationOptions('onboarding.version', queryClient, { silent: true }),
-  );
-  const saveProfileName = useMutation(
-    saveSettingMutationOptions('profile.name', queryClient, { silent: true }),
-  );
-  const saveProfileTimezone = useMutation(
-    saveSettingMutationOptions('profile.timezone', queryClient, { silent: true }),
-  );
-
-  const onboardingStatus = settings?.['onboarding.status'];
-  const onboardingVersion = settings?.['onboarding.version'] ?? '1';
-  const profileName = settings?.['profile.name']?.trim() ?? '';
-  const profileTimezone = settings?.['profile.timezone']?.trim() ?? '';
-  const hasProfileName = profileName.length > 0;
-  const hasProfileTimezone = profileTimezone.length > 0;
-  const hasEnabledProvider = (providers ?? []).some((provider) => provider.enabled);
-  const memoryEnabled = settings?.['memory.enabled'] === 'true';
-  const hasMemoryModelConfigured =
-    (settings?.['memory.embedding.providerId']?.trim().length ?? 0) > 0 &&
-    (settings?.['memory.embedding.modelId']?.trim().length ?? 0) > 0;
-  const hasMemoryPreference = !memoryEnabled || hasMemoryModelConfigured;
-  const isOnboarded = onboardingStatus === 'completed';
-  const isLatestOnboardingVersion = onboardingVersion === CURRENT_ONBOARDING_VERSION;
-  const isOnboardingComplete =
-    isOnboarded &&
-    isLatestOnboardingVersion &&
-    hasProfileName &&
-    hasProfileTimezone &&
-    hasEnabledProvider &&
-    hasMemoryPreference;
-
-  const completeOnboarding = React.useCallback(async () => {
-    await saveOnboardingStatus.mutateAsync('completed');
-    await saveOnboardingVersion.mutateAsync(CURRENT_ONBOARDING_VERSION);
-  }, [saveOnboardingStatus, saveOnboardingVersion]);
-
-  React.useEffect(() => {
-    if (step !== 'success') return;
-
-    const timeout = window.setTimeout(() => {
-      setDismissed(true);
-    }, SUCCESS_CLOSE_DELAY_MS);
-
-    return () => {
-      window.clearTimeout(timeout);
-    };
-  }, [step]);
-
-  const isLoading = isSettingsPending || isProvidersPending;
-  const open = !isLoading && !dismissed && !isOnboardingComplete;
-
-  if (!open) {
+  if (state.isLoading || state.dismissed || state.isOnboardingComplete) {
     return null;
   }
 
@@ -652,88 +24,29 @@ export function OnboardingDialog() {
         showCloseButton={false}
       >
         <div className="flex h-full flex-col p-8">
-          {step === 'welcome' && (
-            <div className="flex h-full flex-col items-center justify-center gap-6 text-center">
-              <div className="flex size-12 items-center justify-center rounded-2xl bg-primary/10 text-primary">
-                <SparklesIcon className="size-6" />
-              </div>
-              <div className="max-w-lg space-y-2">
-                <h2 className="text-2xl font-semibold tracking-tight">Welcome to Stitch</h2>
-                <p className="text-sm text-muted-foreground">
-                  Let&apos;s personalize your profile and connect your first provider so you can
-                  start chatting in less than a minute.
-                </p>
-              </div>
-              <Button size="lg" onClick={() => setStep('profile')}>
-                Continue
-              </Button>
-            </div>
-          )}
+          {state.step === 'welcome' && <WelcomeStep onContinue={() => state.goToStep('profile')} />}
 
-          {step === 'profile' && (
-            <OnboardingProfileStep
-              initialName={profileName}
-              initialTimezone={profileTimezone}
-              isSaving={
-                saveProfileName.isPending ||
-                saveProfileTimezone.isPending ||
-                saveOnboardingStatus.isPending ||
-                saveOnboardingVersion.isPending
-              }
-              onContinue={(name, timezone) => {
-                void Promise.all([
-                  saveProfileName.mutateAsync(name),
-                  saveProfileTimezone.mutateAsync(timezone),
-                ])
-                  .then(() => {
-                    if (hasEnabledProvider) {
-                      setStep('memory');
-                      return undefined;
-                    }
-                    setStep('provider');
-                    return undefined;
-                  })
-                  .catch(() => undefined);
-              }}
+          {state.step === 'profile' && (
+            <ProfileStep
+              initialName={state.profileName}
+              initialTimezone={state.profileTimezone}
+              isSaving={state.isSavingProfile}
+              onContinue={state.saveProfileAndAdvance}
             />
           )}
 
-          {step === 'provider' && (
-            <OnboardingProviderStep
-              onConnected={() => {
-                setStep('memory');
-              }}
+          {state.step === 'provider' && (
+            <ProviderStep onConnected={() => state.goToStep('memory')} />
+          )}
+
+          {state.step === 'memory' && (
+            <MemoryStep
+              onComplete={state.completeOnboarding}
+              onBackToProviders={() => state.goToStep('provider')}
             />
           )}
 
-          {step === 'memory' && (
-            <OnboardingMemoryStep
-              onComplete={() => {
-                void completeOnboarding()
-                  .then(() => {
-                    setStep('success');
-                  })
-                  .catch(() => undefined);
-              }}
-              onBackToProviders={() => {
-                setStep('provider');
-              }}
-            />
-          )}
-
-          {step === 'success' && (
-            <div className="flex h-full flex-col items-center justify-center gap-4 text-center">
-              <div className="flex size-14 items-center justify-center rounded-full bg-emerald-500/10 text-emerald-600">
-                <CheckCircle2Icon className="size-8" />
-              </div>
-              <div className="space-y-1">
-                <h2 className="text-xl font-semibold">You&apos;re all set</h2>
-                <p className="text-sm text-muted-foreground">
-                  Setup complete. Launching your workspace...
-                </p>
-              </div>
-            </div>
-          )}
+          {state.step === 'success' && <SuccessStep />}
         </div>
       </DialogContent>
     </Dialog>
